requirements.txt
# Core ML Libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# NLP & Text Processing
nltk>=3.8
spacy>=3.5.0
sentencepiece>=0.1.99

# Continual Learning Specific
avalanche-lib>=0.4.0

# Active Learning
modAL>=0.4.1

# API & Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Data Storage & Processing
sqlalchemy>=2.0.0
redis>=4.6.0
psycopg2-binary>=2.9.0

# Monitoring & Logging
wandb>=0.15.0
tensorboard>=2.13.0
loguru>=0.7.0

# Evaluation & Metrics
scipy>=1.11.0
seaborn>=0.12.0
matplotlib>=3.7.0

# Privacy & Security
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
cryptography>=41.0.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0
